% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblingdim.R
\name{roblingdim}
\alias{roblingdim}
\title{roblingdim}
\usage{
roblingdim(picname)
}
\arguments{
\item{picname}{Name of the picture}
}
\value{
This returns the dimensions of the image
}
\description{
roblingdim
}
\examples{
roblingdim()
roblingdim(monalisa_jpg)
}
