% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblinghist.R
\name{roblinghist}
\alias{roblinghist}
\title{roblinghist}
\usage{
roblinghist(picname)
}
\arguments{
\item{picname}{Name of the picture}
}
\value{
This will return a histogram of the RGB frequencies
}
\description{
roblinghist
}
\examples{
roblinghist()
roblinghist(monalisa_jpg)
}
