% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblingpixelate.R
\name{roblingpixelate}
\alias{roblingpixelate}
\title{roblingpixelate}
\usage{
roblingpixelate(picname)
}
\arguments{
\item{picname}{Name of the picture}
}
\value{
This returns a pixealted image.
}
\description{
roblingpixelate
}
\examples{
roblingpixelate()
roblingpixelate(monalisa_jpg)
}
