% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblingprint.R
\name{roblingprint}
\alias{roblingprint}
\title{roblingprint}
\usage{
roblingprint(picname)
}
\arguments{
\item{picname}{Name of the the picture}
}
\value{
a picture is printed
}
\description{
roblingprint
}
\examples{
roblingprint()
roblingprint(monalisa_jpg)
}
