% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roblingrgb.R
\name{roblingrgb}
\alias{roblingrgb}
\title{roblingrgb}
\usage{
roblingrgb(picname)
}
\arguments{
\item{picname}{Name of the the picture}
}
\value{
This returns a RGB score chart
}
\description{
roblingrgb
}
\examples{
roblingrgb()
roblingrgb(monalisa_jpg)
}
